records/requirement.txt
django==1.6.5
eav-django